Refactor according to feedback from previous open PRs
 * add documentation and examples
 * remove unused configurations and rename to timeoutConf
 * rename the RemoteContext
 * move generic.chain RpcHandler and DataFetcher to a generic.rpc package
