+++ src/main/resources/reference.conf
@@ -42,8 +42,6 @@ logger.scala.slick = INFO
 
 
 #Used for configuring time-out for reading a web response content
-#Timeout on sync network calls
-requestAwaitTime: 120 s
 #Timeouts to read the response into a strict string
 GETResponseEntityTimeout: 90 s
 POSTResponseEntityTimeout: 1 s
