+++ src/main/scala/tech/cryptonomic/conseil/generic/rpc/RpcHandler.scala
@@ -1,14 +1,18 @@
-package tech.cryptonomic.conseil.generic.chain
+package tech.cryptonomic.conseil.generic.rpc
 
 import cats._
 import cats.arrow._
 import cats.data.Kleisli
 
 /** Defines operations to execute calls against a remote node
-  * @tparam Eff the effect associated with execution (Future, IO, Either, Id...)
-  * @tparam Command represents the input to identify the endpoint (e.g. a path string)
-  * @tparam Reponse represents the output from the node (e.g. a json string)
-  */
+ *
+ * For the sake of comprehension, we can compare a Kleisli[F, A, B] data-type
+ * as a "special" function that wraps the result in some "effect" F, i.e. A => F[B]
+ *
+ * @tparam Eff the effect associated with execution (Future, IO, Either, Id...)
+ * @tparam Command represents the input to identify the endpoint (e.g. a path string)
+ * @tparam Reponse represents the output from the node (e.g. a json string)
+ */
 trait RpcHandler[Eff[_], Command, Response] {
   import cats.data.Kleisli
 
@@ -49,7 +53,22 @@ object RpcHandler {
   ): Eff[Response] =
     rpc.postQuery.run((command, payload))
 
-  /** provides an implicit to convert the effect-type of the handler, if there's a "natural transformation" `F ~> G` in scope */
+  /** Provides a generic way to convert the effect-type of an rpc handler, if there's
+    * a "natural transformation" `F ~> G` in scope
+    * Use case
+    * - suppose we have a RpcHandler using `Future`s as the effect, e.g. RpcHandler[Future, String, String]
+    * - suppose also that we can convert every Future[A] to a IO[A], independently of its content type `A`.
+    *   This is done defining a "Higher-Kinded Function" `FunctionK`, also commonly referred to as
+    *   a "natural transformation" between functors `Future` and `IO`, and denoted as `Future ~> IO`.
+    *
+    * We have here defined an additional transformation, based on `Future ~> IO` that can be used to convert
+    * our RpcHandler[Future, String, String] to a RpcHandler[IO, String, String], in a natural and
+    * purely mechanical way, with no need to manually implement the latter.
+    *
+    * For the sake of completeness, this specific example can be misleading, as using this transformation
+    * for the Future ~> IO case is inefficient and technically wrong, because of the "strict" nature of Future's
+    * evaluation strategy, as opposed to IO's "lazy" evaluation.
+    */
   def functionK[F[_], G[_], Command, Payload](implicit nat: F ~> G) =
     new FunctionK[Aux[F, Command, ?, Payload], Aux[G, Command, ?, Payload]] {
